\name{msk.macrophytes.2017_RiverTypes_DefLimitsUnc}

\alias{msk.macrophytes.2017_RiverTypes_DefLimitsUnc}

\title{
Data frame containing the definition of uncertainty of attribute limits of river types.
}

\description{
Data frame containing the definition of uncertainty of attribute limits of river types.
}

