\name{msk.macrophytes.2017_RiverTypes_DefObsUnc}

\alias{msk.macrophytes.2017_RiverTypes_DefObsUnc}

\title{
Data frame containing the definition of observation uncertainty for river types.
}

\description{
Data frame containing the definition of observation uncertainty for river types.
}

